Refactor storage class into package
